Library Chooser
As a college student, it's troublesome to check library websites every time. This makes up my mind to make a small program to do so...
This is currently only based on the closing time. And only for Columbia libraries.
After adding ADI Density's data of people density in some libraries, this will be better. 
But whether I'll make this change is dependent on my schedule... So hopefully I will get enough free time.....(x